backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
botocore==1.34.158
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==4.4.2
distro==1.9.0
elevenlabs==1.5.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jiter==0.5.0
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
moviepy==1.0.3
nest-asyncio==1.6.0
numpy==2.0.1
openai==1.40.3
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyzmq==26.1.0
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sseclient==0.0.27
stack-data==0.6.3
starlette==0.37.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
twelvelabs==0.2.6
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wcwidth==0.2.13
websockets==12.0
